report: fix first module summary generation
